Updates to republish the plugin on upgrade.
git-svn-id: https://plugins.svn.wordpress.org/optima-express/trunk@445885 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ iHomefinderInstaller.php
@@ -64,7 +64,7 @@ if(!class_exists('IHomefinderInstaller')){
 			
 			$currentVersion=get_option(IHomefinderConstants::VERSION_OPTION);
 			
-			if( $currentVersion != IHomefinderConstants::VERSION ){
+			if( $currentVersion != IHomefinderConstants::VERSION && $this->ihfAdmin->previouslyActivated() ){
 				$this->ihfAdmin->deleteAuthenticationToken() ;
 				$this->ihfAdmin->updateAuthenticationToken() ;
 				$this->ihfRewriteRules->initialize();
